getIndexOfChild(Place*); DeletePlacesView update
There are two getIndexOfChild methods, one that takes a string and
another that takes a pointer. Turns out that comparing points is as easy
as == .
(http://stackoverflow.com/questions/9086372/how-to-compare-pointers)

DeletePlacesView updated to use the new code (Old code commented
inline).

Commented code to be removed, it's only there for review.